+++ backend/api/shops/serializers/ProductSerializer.py
@@ -105,11 +105,11 @@ class SubCategorySerializerCustom(serializers.ModelSerializer):
 		fields = "id","name"
 
 class ControlProductSerializer(serializers.Serializer):
-	quantity = serializers.IntegerField(required=True)
+	quantity = serializers.FloatField(required=True)
 
 
 class SupplyProductSerializer(serializers.Serializer):
-	quantity = serializers.IntegerField(required=True)
+	quantity = serializers.FloatField(required=True)
 	total_buy_price = serializers.FloatField(required=True)
 	sale_price = serializers.FloatField(required=False, allow_null=True)
 	created_at = serializers.DateTimeField(required=False, allow_null=True)
